feat: useTeams hook
