+++ packages/platform/atoms/hooks/event-types/public/useTeamEventTypes.ts
@@ -22,5 +22,6 @@ export const useTeamEventTypes = (teamId: number) => {
         throw new Error(res.data.error.message);
       });
     },
+    enabled: !!orgId && !!teamId,
   });
 };
