+++ packages/platform/types/index.ts
@@ -7,3 +7,4 @@ export * from "./bookings";
 export * from "./billing";
 export * from "./schedules";
 export * from "./event-types";
+export * from "./organizations";
