+++ packages/platform/types/organizations/index.ts
@@ -1,0 +1,1 @@
+export * from "./teams";
