+++ packages/platform/types/organizations/teams/index.ts
@@ -1,0 +1,1 @@
+export * from "./outputs";
