+++ packages/platform/types/organizations/teams/outputs/index.ts
@@ -1,0 +1,1 @@
+export * from "./team.output";
